convolving and pooling

+++ math/convolutions_and_pooling/2-convolve_grayscale_padding.py
@@ -19,7 +19,7 @@ def convolve_grayscale_padding(images, kernel, padding):
     images_padded = np.pad(images, ((0, 0), (ph, ph), (pw, pw)), mode='constant')
     output_h = h + 2 * ph - kh + 1
     output_w = w + 2 * pw - kw + 1
-    output = np.zeros((m, output_h, output_w), dtype=np.int32)
+    output = np.zeros((m, output_h, output_w))
 
     for i in range(output_h):
         for j in range(output_w):
